[Logs] Deprecate configuration settings (#201625)
## 📓 Summary

Closes #200898

These changes deprecate some unused configurations and update the
implementation where required in preparation for the Kibana v9 upgrade.

<img width="3004" alt="Screenshot 2024-11-25 at 12 54 14"
src="https://github.com/user-attachments/assets/cfa56d25-a270-4ec5-a97a-e72e7a7478a4">

---------

(cherry picked from commit 2994b0001cbae6bc1528bc1ad77c435597e0d5a2)

# Conflicts:
#	x-pack/plugins/observability_solution/infra/tsconfig.json
