+++ x-pack/plugins/observability_solution/infra/tsconfig.json
@@ -116,7 +116,10 @@
     "@kbn/entityManager-plugin",
     "@kbn/entities-schema",
     "@kbn/zod",
-    "@kbn/observability-utils-server"
+    "@kbn/observability-utils-server",
+    "@kbn/core-plugins-server",
+    "@kbn/config",
+    "@kbn/observability-utils-common"
   ],
   "exclude": ["target/**/*"]
 }
