+++ x-pack/plugins/observability_solution/logs_shared/server/lib/logs_shared_types.ts
@@ -23,4 +23,5 @@ export interface LogsSharedBackendLibs extends LogsSharedDomainLibs {
   getStartServices: LogsSharedPluginStartServicesAccessor;
   getUsageCollector: () => UsageCollector;
   logger: Logger;
+  isServerless: boolean;
 }
