+++ x-pack/plugins/observability_solution/observability_onboarding/public/application/shared/header_action_menu.tsx
@@ -22,15 +22,14 @@ interface Props {
 
 export function ObservabilityOnboardingHeaderActionMenu({ setHeaderActionMenu, theme$ }: Props) {
   const {
-    services: { config },
+    services: { context },
   } = useKibana<ObservabilityOnboardingAppServices>();
   const location = useLocation();
   const normalizedPathname = location.pathname.replace(/\/$/, '');
 
   const isRootPage = normalizedPathname === '';
-  const isServerless = config.serverless.enabled;
 
-  if (!isServerless && !isRootPage) {
+  if (!context.isServerless && !isRootPage) {
     return (
       <HeaderMenuPortal setHeaderActionMenu={setHeaderActionMenu} theme$={theme$}>
         <EuiButton
