+++ x-pack/plugins/observability_solution/observability_onboarding/tsconfig.json
@@ -43,7 +43,8 @@
     "@kbn/deeplinks-analytics",
     "@kbn/custom-integrations-plugin",
     "@kbn/server-route-repository-utils",
-    "@kbn/core-application-browser"
+    "@kbn/core-application-browser",
+    "@kbn/core-plugins-server"
   ],
   "exclude": [
     "target/**/*"
